+++ apps/core/src/components/affine/new-workspace-setting-detail/publish.tsx
@@ -17,8 +17,8 @@ import { useCallback, useEffect, useMemo, useState } from 'react';
 import { toast } from '../../../utils';
 import { EnableAffineCloudModal } from '../enable-affine-cloud-modal';
 import { TmpDisableAffineCloudModal } from '../tmp-disable-affine-cloud-modal';
-import type { WorkspaceSettingDetailProps } from './index';
 import * as style from './style.css';
+import type { WorkspaceSettingDetailProps } from './types';
 
 export interface PublishPanelProps
   extends Omit<WorkspaceSettingDetailProps, 'workspaceId'> {
